// src/common/types/enums.ts
export enum PaymentStatus {
  SUCCESS = 'success',
  PENDING = 'pending',
  FAILED = 'failed',
}
export enum RegistationStatus {
  COMPLETED = 'completed',
  PENDING = 'pending',
  FAILED = 'failed',
}

export enum RoleEnum{
  ADMIN = 'admin',
  USER = 'user',
  ORGANIZER = 'organizer',
  GUEST = 'guest'
}